changed location of addition added to custom propType and deleted previous location
