+++ src/_playground/documentation/Properties/_PropertyDefault.js
@@ -17,7 +17,7 @@ const PropertyDefault = ({ defaultValue, prop }) => {
     if (prop.typeName === 'i18n') {
         const list = Object.keys(defaultValue).map((t, i) => {
             return (
-                <div key={i}>{`${t}: ${defaultValue[t]}`}</div>);
+                <div key={i}>{`${t}: '${defaultValue[t]}'`}</div>);
         });
         return (
             <React.Fragment>
